requirements.txt
# Core Framework
crewai==0.70.1
crewai-tools==0.12.1
streamlit==1.39.0
fastapi==0.115.5
uvicorn==0.32.0

# LLM and AI
openai==1.54.5
# langchain==0.3.7
# langchain-openai==0.2.8
# langchain-community==0.3.7

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4
pypandoc==1.14
reportlab==4.2.5

# Firebase
firebase-admin==6.6.0

# Vector Database
chromadb==0.4.24

# Web Research Tools
requests==2.32.3
beautifulsoup4==4.12.3
serpapi==0.1.5

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Additional Tools
aiofiles==24.1.0
python-multipart==0.0.17
